+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -18,7 +18,7 @@
 #include "EnumExplicitCheck.h"
 #include "EnumInitCheck.h"
 #include "EnumScopedCheck.h"
-#include "ForLoopCheck.h"
+#include "ForLoopCounterCheck.h"
 #include "FriendDeclCheck.h"
 #include "FunctionNameUseCheck.h"
 #include "LambdaImplicitCaptureCheck.h"
@@ -74,8 +74,8 @@ public:
         "bsl-enum-init");
     CheckFactories.registerCheck<EnumScopedCheck>(
         "bsl-enum-scoped");
-    CheckFactories.registerCheck<ForLoopCheck>(
-        "bsl-for-loop");
+    CheckFactories.registerCheck<ForLoopCounterCheck>(
+        "bsl-for-loop-counter");
     CheckFactories.registerCheck<FriendDeclCheck>(
         "bsl-friend-decl");
     CheckFactories.registerCheck<FunctionNameUseCheck>(